Buffer changed to long generate Scanner fixed makefile

+++ sysprogWS2012/Buffer/src/Buffer.cpp
@@ -35,11 +35,11 @@ char Buffer::getNextChar() {
 	return input_->getNextChar();
 }
 
-int Buffer::getCurrentRow() {
+long Buffer::getCurrentRow() {
 	return input_->getCurrentRow();
 }
 
-int Buffer::getCurrentColumn() {
+long Buffer::getCurrentColumn() {
 	return input_->getCurrentColumn();
 }
